requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
importlib-metadata==4.2.0
mccabe==0.7.0
Pillow==9.3.0
prettytable==3.5.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyTelegramBotAPI==4.7.1
python-dotenv==0.21.0
requests==2.28.1
typing-extensions==4.4.0
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.10.0